pyproject.toml
[tool.poetry]
name = "whatstandard"
version = "0.1.0"
description = "This is a chatbot to ask a consensus of a treatment based on pooled guidelines of a specific disorder"
authors = ["Toshiki"]
license = "Mozilla Public License 2.0"
readme = "README.md"
[tool.poetry.group.dev]

[tool.poetry.dependencies]
python = "^3.12"
langchain = "*"
langchainhub = "*"
langchain-chroma ="*"
langchain-groq = "*"
langchain-huggingface = "*"
langchain_community = "*"
gradio = "*"
pypdf = "*"
rapidocr-onnxruntime = "*"
pypdf2 = "^3.0.1"
tokenizers = "^0.20.0"
pandas = "^2.2.3"
jupyter-contrib-nbextensions = "^0.7.0"
rank-bm25 = "^0.2.2"
sentence-transformers = "^3.1.1"
scipy = "^1.14.1"
faiss-cpu = "^1.9.0"
[tool.poetry.group.dev.dependencies]
jupyter = "*"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
